+++ mobile/.eslintignore
@@ -3,7 +3,7 @@ node_modules/
 npm-debug.*
 web-build/
 coverage
-src/.common
+src/_common
 
 # ignorin native code
 android/
