+++ mobile/package.json
@@ -8,7 +8,7 @@
     "url": "https://omarbelghaouti.space/"
   },
   "scripts": {
-    "common:sync": "rimraf ./src/.common && syncdir ../common/src ./src/.common -c",
+    "common:sync": "rimraf ./src/_common && syncdir ../common/src ./src/_common -c",
     "common:sync:watch": "yarn common:sync --watch",
     "expo": "expo",
     "dev": "expo start",
