+++ mobile/src/config/index.ts
@@ -1,4 +1,4 @@
-import { fsConfig } from "../.common/config";
+import { fsConfig } from "../_common/config";
 import { getEnv } from "../utils/env";
 
 export const fullstackConfig = fsConfig(getEnv(), { hostname: process.env.LOCAL_API_HOST });
